minor changes to forms
