--- frontend/src/components/formsContainer/FormGenerator.css
@@ -1,13 +1,0 @@
-/* FormsContainer.css */
-.forms-container {
-  background-color: #f1f1f1;
-  padding: 20px;
-  border-radius: 4px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
