--- frontend/src/components/formsContainer/FormGenerator.js
@@ -1,33 +1,0 @@
-import React, { useState } from 'react';
-
-const Form = ({ dropdownItems, selectionField }) => {
-  const [selectedValue, setSelectedValue] = useState('');
-
-  const handleDropdownChange = (event) => {
-    setSelectedValue(event.target.value);
-  };
-
-  const clearSelection = () => {
-    setSelectedValue('');
-  };
-
-  return (
-    <form>
-      <label htmlFor={selectionField}>Select {selectionField}:</label>
-      <select id={selectionField} value={selectedValue} onChange={handleDropdownChange}>
-        <option value="">-- Select --</option>
-        {dropdownItems.map((item) => (
-          <option key={item} value={item}>
-            {item}
-          </option>
-        ))}
-      </select>
-
-      {selectedValue && (
-        <p onClick={clearSelection}>Selected: {selectedValue}</p>
-      )}
-    </form>
-  );
-};
-
-export default Form;
